What a mess, but scenes now work.

+++ practical_4/ghost.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include "entity.h"
 
 class Ghost : public Entity {
@@ -6,7 +7,7 @@ private:
 	float _speed;
 
 public:
-	void update(double dt) override;
+	void update(float dt) override;
 	Ghost();
 	void render(sf::RenderWindow &window) const override;
 };
