+++ practical_4/player.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include "entity.h"
 
 class Player : public Entity {
@@ -6,7 +7,7 @@ private:
 	float _speed;
 
 public:
-	void update(double dt) override;
+	void update(float dt) override;
 	Player();
 	void render(sf::RenderWindow &window) const override;
 };
